+++ src/module.ts
@@ -21,9 +21,25 @@
  * limitations under the License.
  */
 
-import { bridgedNode, MatterbridgeDynamicPlatform, MatterbridgeEndpoint, onOffLight, onOffOutlet, onOffSwitch, PlatformConfig, PlatformMatterbridge } from 'matterbridge';
-import { isValidObject } from 'matterbridge/utils';
-import { AnsiLogger } from 'matterbridge/logger';
+import {
+  bridgedNode,
+  colorTemperatureLight,
+  CommandHandlerData,
+  dimmableLight,
+  extendedColorLight,
+  MatterbridgeColorControlServer,
+  MatterbridgeDynamicPlatform,
+  MatterbridgeEndpoint,
+  MatterbridgeLevelControlServer,
+  onOffLight,
+  onOffOutlet,
+  onOffSwitch,
+  PlatformConfig,
+  PlatformMatterbridge,
+} from 'matterbridge';
+import { hslColorToRgbColor, isValidNumber, isValidObject, isValidString, miredToKelvin, wait } from 'matterbridge/utils';
+import { AnsiLogger, rs } from 'matterbridge/logger';
+import { ColorControl, LevelControl } from 'matterbridge/matter/clusters';
 
 import { fetch } from './fetch.js';
 
@@ -32,12 +48,30 @@ export interface WebhookConfig {
   httpUrl: string;
   test: boolean;
 }
+export interface OutletConfig {
+  onUrl: string;
+  offUrl: string;
+}
+
+export interface LightConfig {
+  minMireds: number;
+  maxMireds: number;
+  onUrl: string;
+  offUrl: string;
+  brightnessUrl: string;
+  colorTempUrl: string;
+  rgbUrl: string;
+}
 
 export type WebhooksPlatformConfig = PlatformConfig & {
   whiteList: string[];
   blackList: string[];
+  // Normal webhooks device type
   deviceType: 'Outlet' | 'Switch' | 'Light';
   webhooks: Record<string, WebhookConfig>;
+  // Devices configurations
+  outlets: Record<string, OutletConfig>;
+  lights: Record<string, LightConfig>;
 };
 
 /**
@@ -54,10 +88,11 @@ export default function initializePlugin(matterbridge: PlatformMatterbridge, log
 }
 
 export class WebhooksPlatform extends MatterbridgeDynamicPlatform {
-  private webhooks: Record<string, WebhookConfig>;
-  readonly bridgedDevices = new Map<string, MatterbridgeEndpoint>();
-
-  constructor(matterbridge: PlatformMatterbridge, log: AnsiLogger, config: WebhooksPlatformConfig) {
+  constructor(
+    matterbridge: PlatformMatterbridge,
+    log: AnsiLogger,
+    override config: WebhooksPlatformConfig,
+  ) {
     super(matterbridge, log, config);
 
     // Verify that Matterbridge is the correct version
@@ -66,8 +101,6 @@ export class WebhooksPlatform extends MatterbridgeDynamicPlatform {
     }
 
     this.log.info('Initializing platform:', this.config.name);
-
-    this.webhooks = config.webhooks;
 
     this.log.info('Finished initializing platform:', this.config.name);
   }
@@ -79,19 +112,19 @@ export class WebhooksPlatform extends MatterbridgeDynamicPlatform {
     await this.ready;
     await this.clearSelect();
 
-    // Register devices
-    let i = 0;
-    for (const webhookName in this.webhooks) {
-      this.log.debug(`Loading webhook ${++i} ${webhookName} with method ${this.webhooks[webhookName].method} and url ${this.webhooks[webhookName].httpUrl}`);
+    // Register webhooks devices (deviceName: webhookName, serialNumber: webhook+i)
+    let i = 1;
+    for (const webhookName in this.config.webhooks) {
+      this.log.debug(`Loading webhook ${i} ${webhookName} with method ${this.config.webhooks[webhookName].method} and url ${this.config.webhooks[webhookName].httpUrl}`);
 
-      const webhook = this.webhooks[webhookName];
+      const webhook = this.config.webhooks[webhookName];
       this.setSelectDevice('webhook' + i, webhookName, undefined, 'hub');
       if (!this.validateDevice(['webhook' + i, webhookName], true)) continue;
       this.log.info(`Registering device: ${webhookName} with method ${webhook.method} and url ${webhook.httpUrl}`);
       const device = new MatterbridgeEndpoint(
         [this.config.deviceType === 'Outlet' ? onOffOutlet : this.config.deviceType === 'Light' ? onOffLight : onOffSwitch, bridgedNode],
         { id: webhookName },
-        this.config.debug as boolean,
+        this.config.debug,
       )
         .createDefaultBridgedDeviceBasicInformationClusterServer(
           webhookName,
@@ -100,13 +133,14 @@ export class WebhooksPlatform extends MatterbridgeDynamicPlatform {
           'Matterbridge',
           'Matterbridge Webhook',
           0,
-          this.config.version as string,
+          this.config.version,
         )
         .createOnOffClusterServer(false)
         .addRequiredClusterServers()
         .addCommandHandler('on', async () => {
-          this.log.info(`Webhook ${webhookName} triggered.`);
+          this.log.info(`Webhook ${webhookName} triggered`);
           await device.setAttribute('onOff', 'onOff', false, device.log);
+          this.log.debug(`Fetching ${webhook.httpUrl} with ${webhook.method}...`);
           fetch(webhook.httpUrl, webhook.method)
             .then(() => this.log.notice(`Webhook ${webhookName} successful!`))
             .catch((err) => {
@@ -114,17 +148,106 @@ export class WebhooksPlatform extends MatterbridgeDynamicPlatform {
             });
         });
       await this.registerDevice(device);
-      this.bridgedDevices.set(webhookName, device);
+    }
+
+    // Register outlet devices (deviceName: outletName, serialNumber: outlet+i)
+    i = 1;
+    for (const outletName in this.config.outlets) {
+      this.log.debug(`Loading outlet ${i} ${outletName}...`);
+
+      const webhook = this.config.outlets[outletName];
+      this.setSelectDevice('outlet' + i, outletName, undefined, 'hub');
+      if (!this.validateDevice(['outlet' + i, outletName], true)) continue;
+      this.log.info(`Registering outlet: ${outletName}...`);
+      const device = new MatterbridgeEndpoint([onOffOutlet, bridgedNode], { id: outletName }, this.config.debug)
+        .createDefaultBridgedDeviceBasicInformationClusterServer(
+          outletName,
+          'outlet' + i++,
+          this.matterbridge.aggregatorVendorId,
+          'Matterbridge',
+          'Matterbridge Webhook Outlet',
+          0,
+          this.config.version,
+        )
+        .createOnOffClusterServer(false)
+        .addRequiredClusterServers()
+        .addCommandHandler('on', async (data) => {
+          this.parseUrl('outlet', outletName, 'on', webhook.onUrl, data);
+        })
+        .addCommandHandler('off', async (data) => {
+          this.parseUrl('outlet', outletName, 'off', webhook.offUrl, data);
+        });
+      await this.registerDevice(device);
+    }
+
+    // Register light devices (deviceName: lightName, serialNumber: light+i)
+    i = 1;
+    for (const lightName in this.config.lights) {
+      this.log.debug(`Loading light ${i} ${lightName}...`);
+
+      const webhook = this.config.lights[lightName];
+      this.setSelectDevice('light' + i, lightName, undefined, 'hub');
+      if (!this.validateDevice(['light' + i, lightName], true)) continue;
+      this.log.info(`Registering light: ${lightName}...`);
+      let deviceType = onOffLight;
+      if (isValidString(webhook.brightnessUrl, 1)) deviceType = dimmableLight;
+      if (isValidString(webhook.colorTempUrl, 1)) deviceType = colorTemperatureLight;
+      if (isValidString(webhook.rgbUrl, 1)) deviceType = extendedColorLight;
+      const device = new MatterbridgeEndpoint([deviceType, bridgedNode], { id: lightName }, this.config.debug)
+        .createDefaultBridgedDeviceBasicInformationClusterServer(
+          lightName,
+          'light' + i++,
+          this.matterbridge.aggregatorVendorId,
+          'Matterbridge',
+          'Matterbridge Webhook Light',
+          0,
+          this.config.version,
+        )
+        .createOnOffClusterServer(false)
+        .createDefaultColorControlClusterServer(undefined, undefined, undefined, undefined, 250, webhook.minMireds, webhook.maxMireds)
+        .createDefaultLevelControlClusterServer()
+        .addRequiredClusterServers()
+        .addCommandHandler('on', async (data) => {
+          this.parseUrl('light', lightName, 'on', webhook.onUrl, data);
+        })
+        .addCommandHandler('off', async (data) => {
+          this.parseUrl('light', lightName, 'off', webhook.offUrl, data);
+        })
+        .addCommandHandler('moveToLevel', async (data) => {
+          this.parseUrl('light', lightName, 'moveToLevel', webhook.brightnessUrl, data);
+        })
+        .addCommandHandler('moveToLevelWithOnOff', async (data) => {
+          this.parseUrl('light', lightName, 'moveToLevelWithOnOff', webhook.brightnessUrl, data);
+        })
+        .addCommandHandler('moveToColorTemperature', async (data) => {
+          this.parseUrl('light', lightName, 'moveToColorTemperature', webhook.colorTempUrl, data);
+        })
+        .addCommandHandler('moveToHueAndSaturation', async (data) => {
+          this.parseUrl('light', lightName, 'moveToHueAndSaturation', webhook.rgbUrl, data);
+        })
+        .addCommandHandler('moveToHue', async (data) => {
+          this.parseUrl('light', lightName, 'moveToHue', webhook.rgbUrl, data);
+        })
+        .addCommandHandler('moveToSaturation', async (data) => {
+          this.parseUrl('light', lightName, 'moveToSaturation', webhook.rgbUrl, data);
+        })
+        .addCommandHandler('moveToColor', async (data) => {
+          this.parseUrl('light', lightName, 'moveToColor', webhook.rgbUrl, data);
+        });
+      await this.registerDevice(device);
     }
   }
 
   override async onConfigure(): Promise<void> {
     await super.onConfigure();
     this.log.info('onConfigure called');
-    this.bridgedDevices.forEach(async (device) => {
-      this.log.info(`Configuring device: ${device.deviceName}`);
-      await device.setAttribute('onOff', 'onOff', false, device.log);
-    });
+    for (const device of this.getDevices()) {
+      // Turn off the normal webhook devices to avoid confusion. For the other devices leave them as is in the matter storage cause the main attributes persist.
+      if (device.deviceName && device.deviceName in this.config.webhooks) {
+        this.log.info(`Configuring device: ${device.deviceName}`);
+        await device.setAttribute('onOff', 'onOff', false, device.log);
+      }
+    }
   }
 
   override async onAction(action: string, value?: string, id?: string, formData?: PlatformConfig): Promise<void> {
@@ -155,9 +278,9 @@ export class WebhooksPlatform extends MatterbridgeDynamicPlatform {
         return;
       }
       // Test the webhook already confirmed
-      for (const webhookName in this.webhooks) {
-        if (Object.prototype.hasOwnProperty.call(this.webhooks, webhookName)) {
-          const webhook = this.webhooks[webhookName];
+      for (const webhookName in this.config.webhooks) {
+        if (Object.prototype.hasOwnProperty.call(this.config.webhooks, webhookName)) {
+          const webhook = this.config.webhooks[webhookName];
           if (id?.includes(webhookName)) {
             this.log.info(`Testing webhook ${webhookName} method ${webhook.method} url ${webhook.httpUrl}`);
             fetch(webhook.httpUrl, webhook.method)
@@ -178,6 +301,135 @@ export class WebhooksPlatform extends MatterbridgeDynamicPlatform {
     await super.onShutdown(reason);
     this.log.info('onShutdown called with reason:', reason ?? 'none');
     if (this.config.unregisterOnShutdown === true) await this.unregisterAllDevices();
-    this.bridgedDevices.clear();
+  }
+
+  /**
+   * Parse the URL to extract method and URL.
+   *
+   * @param {string} deviceType - The type of the device.
+   * @param {string} deviceName - The name of the device.
+   * @param {string} command - The command to parse.
+   * @param {string} url - The URL to parse.
+   * @param {CommandHandlerData} [data] - The command handler data.
+   * @returns {Promise<{ method: 'POST' | 'GET'; url: string }>} - The parsed method and URL.
+   */
+  async parseUrl(deviceType: string, deviceName: string, command: string, url: string, data: CommandHandlerData): Promise<{ method: 'POST' | 'GET'; url: string }> {
+    this.log.info(`Webhook ${deviceType} ${deviceName} ${command} triggered`);
+    const endpoint = data.endpoint;
+    this.log.debug(`Webhook ${deviceType} ${deviceName} ${command} triggered on endpoint ${endpoint?.deviceName}`);
+
+    // Determine method
+    let method: 'POST' | 'GET' = 'GET';
+    let parsedUrl = url;
+    if (url.startsWith('GET#')) {
+      method = 'GET';
+      parsedUrl = url.replace('GET#', '');
+    } else if (url.startsWith('POST#')) {
+      method = 'POST';
+      parsedUrl = url.replace('POST#', '');
+    }
+
+    // Request based replacements
+    if (parsedUrl.includes('${LEVEL}') && isValidNumber(data.request.level)) {
+      parsedUrl = parsedUrl.replace('${LEVEL}', (data.request as LevelControl.MoveToLevelRequest).level.toString());
+    }
+    if (url.includes('${LEVEL100}') && isValidNumber(data.request.level)) {
+      parsedUrl = parsedUrl.replace('${LEVEL100}', Math.round(((data.request as LevelControl.MoveToLevelRequest).level / 254) * 100).toString());
+    }
+    if (parsedUrl.includes('${KELVIN}') && isValidNumber(data.request.colorTemperatureMireds)) {
+      parsedUrl = parsedUrl.replace('${KELVIN}', Math.round(miredToKelvin((data.request as ColorControl.MoveToColorTemperatureRequest).colorTemperatureMireds)).toString());
+    }
+    if (parsedUrl.includes('${MIRED}') && isValidNumber(data.request.colorTemperatureMireds)) {
+      parsedUrl = parsedUrl.replace('${MIRED}', Math.round((data.request as ColorControl.MoveToColorTemperatureRequest).colorTemperatureMireds).toString());
+    }
+    if (parsedUrl.includes('${COLORX}') && isValidNumber(data.request.colorX, 0, 65279)) {
+      parsedUrl = parsedUrl.replace('${COLORX}', this.roundTo((data.request as ColorControl.MoveToColorRequest).colorX / 65536, 4).toString());
+    }
+    if (parsedUrl.includes('${COLORY}') && isValidNumber(data.request.colorY, 0, 65279)) {
+      parsedUrl = parsedUrl.replace('${COLORY}', this.roundTo((data.request as ColorControl.MoveToColorRequest).colorY / 65536, 4).toString());
+    }
+    if (parsedUrl.includes('${HUE}') && isValidNumber(data.request.hue, 0, 254)) {
+      parsedUrl = parsedUrl.replace('${HUE}', Math.round(((data.request as ColorControl.MoveToHueAndSaturationRequest).hue * 360) / 254).toString());
+    }
+    if (parsedUrl.includes('${SATURATION}') && isValidNumber(data.request.saturation, 0, 254)) {
+      parsedUrl = parsedUrl.replace('${SATURATION}', Math.round(((data.request as ColorControl.MoveToHueAndSaturationRequest).saturation * 100) / 254).toString());
+    }
+
+    // Attributes based replacements
+    if ((parsedUrl.includes('${level}') || parsedUrl.includes('${level100}')) && isValidNumber(data.attributes.currentLevel, 1, 254)) {
+      await wait(100); // Wait a bit to ensure the latest values are written by the command handlers. After the wait the attributes should be updated and the context be closed.
+      data.attributes = endpoint.stateOf(MatterbridgeLevelControlServer);
+      if (isValidNumber(data.attributes.currentLevel, 1, 254)) {
+        if (url.includes('${level}')) parsedUrl = parsedUrl.replace('${level}', data.attributes.currentLevel.toString());
+        if (url.includes('${level100}')) parsedUrl = parsedUrl.replace('${level100}', Math.round((data.attributes.currentLevel / 254) * 100).toString());
+      }
+    }
+    if (
+      (parsedUrl.includes('${mired}') || parsedUrl.includes('${kelvin}')) &&
+      isValidNumber(data.attributes.colorTemperatureMireds, data.attributes.colorTempPhysicalMinMireds as number, data.attributes.colorTempPhysicalMaxMireds as number)
+    ) {
+      await wait(100); // Wait a bit to ensure the latest values are written by the command handlers. After the wait the attributes should be updated and the context be closed.
+      data.attributes = endpoint.stateOf(MatterbridgeColorControlServer);
+      if (isValidNumber(data.attributes.colorTemperatureMireds)) {
+        const kelvin = miredToKelvin(data.attributes.colorTemperatureMireds);
+        this.log.debug(`Attribute colorTemperatureMireds is ${data.attributes.colorTemperatureMireds}, which is ${kelvin}K`);
+        if (url.includes('${mired}')) parsedUrl = parsedUrl.replace('${mired}', data.attributes.colorTemperatureMireds.toString());
+        if (url.includes('${kelvin}')) parsedUrl = parsedUrl.replace('${kelvin}', kelvin.toString());
+      }
+    }
+    if (
+      (parsedUrl.includes('${hue}') || parsedUrl.includes('${saturation}') || parsedUrl.includes('${red}') || parsedUrl.includes('${green}') || parsedUrl.includes('${blue}')) &&
+      isValidNumber(data.attributes.currentHue, 0, 254) &&
+      isValidNumber(data.attributes.currentSaturation, 0, 254)
+    ) {
+      await wait(100); // Wait a bit to ensure the latest values are written by the command handlers. After the wait the attributes should be updated and the context be closed.
+      data.attributes = endpoint.stateOf(MatterbridgeColorControlServer);
+      if (isValidNumber(data.attributes.currentHue, 0, 254) && isValidNumber(data.attributes.currentSaturation, 0, 254)) {
+        const rgb = hslColorToRgbColor((data.attributes.currentHue * 360) / 254, (data.attributes.currentSaturation * 100) / 254, 50);
+        this.log.debug(`Converted hue ${data.attributes.currentHue} and saturation ${data.attributes.currentSaturation} to RGB r: ${rgb.r} g: ${rgb.g} b: ${rgb.b}`);
+        if (url.includes('${hue}')) parsedUrl = parsedUrl.replace('${hue}', Math.round((data.attributes.currentHue * 360) / 254).toString());
+        if (url.includes('${saturation}')) parsedUrl = parsedUrl.replace('${saturation}', Math.round((data.attributes.currentSaturation * 100) / 254).toString());
+        if (url.includes('${red}') && rgb) parsedUrl = parsedUrl.replace('${red}', rgb.r.toString());
+        if (url.includes('${green}') && rgb) parsedUrl = parsedUrl.replace('${green}', rgb.g.toString());
+        if (url.includes('${blue}') && rgb) parsedUrl = parsedUrl.replace('${blue}', rgb.b.toString());
+      }
+    }
+    if (
+      (parsedUrl.includes('${colorX}') || parsedUrl.includes('${colorY}')) &&
+      isValidNumber(data.attributes.currentX, 0, 65279) &&
+      isValidNumber(data.attributes.currentY, 0, 65279)
+    ) {
+      await wait(100); // Wait a bit to ensure the latest values are written by the command handlers. After the wait the attributes should be updated and the context be closed.
+      data.attributes = endpoint.stateOf(MatterbridgeColorControlServer);
+      if (isValidNumber(data.attributes.currentX, 0, 65279) && isValidNumber(data.attributes.currentY, 0, 65279)) {
+        if (url.includes('${colorX}')) parsedUrl = parsedUrl.replace('${colorX}', this.roundTo(data.attributes.currentX / 65536, 4).toString());
+        if (url.includes('${colorY}')) parsedUrl = parsedUrl.replace('${colorY}', this.roundTo(data.attributes.currentY / 65536, 4).toString());
+      }
+    }
+
+    // Execute the fetch
+    this.log.debug(`Fetching ${parsedUrl} with ${method}...`);
+    fetch(parsedUrl, method)
+      .then((response) => {
+        this.log.notice(`Webhook ${deviceType} ${deviceName} ${command} successful!`);
+        this.log.debug(`Webhook ${deviceType} ${deviceName} ${command} response:${rs}\n`, response);
+        return;
+      })
+      .catch((err) => {
+        this.log.error(`Webhook ${deviceType} ${deviceName} ${command} failed: ${err instanceof Error ? err.message : err}`);
+      });
+    return { method, url: parsedUrl };
+  }
+
+  /**
+   * Round a number to the number of digits specified
+   *
+   * @param {number} value - The number to round
+   * @param {number} digits - The number of digits to round to
+   * @returns {number} The rounded number
+   */
+  private roundTo(value: number, digits: number): number {
+    const factor = Math.pow(10, digits);
+    return Math.round(value * factor) / factor;
   }
 }